requirements.txt
argparse
requests
pycryptodomex
